feat: modify users/tweets seed file
