+++ seeders/20210707064422-users-seed-file.js
@@ -1,76 +1,32 @@
-'use strict'
+'use strict';
 const bcrypt = require('bcryptjs')
+const users = Array.from({ length: 5 }).map((d, i) => ({
+  account: `user${i + 1}`,
+  name: `user${i + 1}`,
+  email: `user${i + 1}@example.com`,
+  password: bcrypt.hashSync('12345678', bcrypt.genSaltSync(10), null),
+  is_admin: false,
+  img: `https://loremflickr.com/250/250/avator/?random=${i}`,
+  createdAt: new Date(),
+  updatedAt: new Date()
+}))
+users.unshift({
+  account: 'root',
+  name: 'root',
+  email: 'root@example.com',
+  password: bcrypt.hashSync('12345678', bcrypt.genSaltSync(10), null),
+  is_admin: true,
+  img: 'https://loremflickr.com/250/250/avator/?random=50',
+  createdAt: new Date(),
+  updatedAt: new Date()
+})
+
 module.exports = {
   up: async (queryInterface, Sequelize) => {
-    await queryInterface.bulkInsert('Users', [{
-      account: 'root',
-      email: 'root@example.com',
-      password: bcrypt.hashSync('12345678', bcrypt.genSaltSync(10), null),
-      description: 'admin account',
-      img: 'https://loremflickr.com/250/250/avator/?random=1',
-      bg_img: '',
-      is_admin: true,
-      name: 'root',
-      createdAt: new Date(),
-      updatedAt: new Date()
-    }, {
-      account: 'user1',
-      email: 'user1@example.com',
-      password: bcrypt.hashSync('12345678', bcrypt.genSaltSync(10), null),
-      description: 'user account',
-      img: 'https://loremflickr.com/250/250/avator/?random=3',
-      bg_img: '',
-      is_admin: false,
-      name: 'user1',
-      createdAt: new Date(),
-      updatedAt: new Date()
-    }, {
-      account: 'user2',
-      email: 'user2@example.com',
-      password: bcrypt.hashSync('12345678', bcrypt.genSaltSync(10), null),
-      description: 'user account',
-      img: 'https://loremflickr.com/250/250/avator/?random=5',
-      bg_img: '',
-      is_admin: false,
-      name: 'user2',
-      createdAt: new Date(),
-      updatedAt: new Date()
-    }, {
-      account: 'user3',
-      email: 'user3@example.com',
-      password: bcrypt.hashSync('12345678', bcrypt.genSaltSync(10), null),
-      description: 'user account',
-      img: 'https://loremflickr.com/250/250/avator/?random=7',
-      bg_img: '',
-      is_admin: false,
-      name: 'user3',
-      createdAt: new Date(),
-      updatedAt: new Date()
-    }, {
-      account: 'user4',
-      email: 'user4@example.com',
-      password: bcrypt.hashSync('12345678', bcrypt.genSaltSync(10), null),
-      description: 'user account',
-      img: 'https://loremflickr.com/250/250/avator/?random=9',
-      bg_img: '',
-      is_admin: false,
-      name: 'user4',
-      createdAt: new Date(),
-      updatedAt: new Date()
-    }, {
-      account: 'user5',
-      email: 'user5@example.com',
-      password: bcrypt.hashSync('12345678', bcrypt.genSaltSync(10), null),
-      description: 'user account',
-      img: 'https://loremflickr.com/250/250/avator/?random=11',
-      bg_img: '',
-      is_admin: false,
-      name: 'user5',
-      createdAt: new Date(),
-      updatedAt: new Date()
-    }], {})
+    await queryInterface.bulkInsert('Users', users, {})
   },
+
   down: async (queryInterface, Sequelize) => {
     await queryInterface.bulkDelete('Users', null, {})
   }
-}
+};
