+++ seeders/20210707065720-tweets-seed-file.js
@@ -5,13 +5,13 @@ module.exports = {
   up: async (queryInterface, Sequelize) => {
     await queryInterface.bulkInsert('Tweets',
       Array.from({ length: 50 }).map((d, i) =>
-        ({
-          UserId: Math.floor(Math.random() * 5) + 1,
-          content: faker.lorem.text(),
-          likes: 0,
-          createdAt: new Date(),
-          updatedAt: new Date()
-        })
+      ({
+        UserId: Math.floor(i / 10) + 2,
+        content: faker.lorem.text(),
+        likes: 0,
+        createdAt: new Date(),
+        updatedAt: new Date()
+      })
       ), {})
   },
