--- seeders/20210707185602-tweets-seed-file.js
@@ -1,17 +1,0 @@
-'use strict';
-const faker = require('faker')
-module.exports = {
-  up: async (queryInterface, Sequelize) => {
-    await queryInterface.bulkInsert('Tweets', Array.from({ length: 50 }).map((d, i) => ({
-      UserId: Math.floor(Math.random() * 5) + 1,
-      content: faker.lorem.text(),
-      likes: 0,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    })), {})
-  },
-
-  down: async (queryInterface, Sequelize) => {
-    await queryInterface.bulkDelete('Tweets', null, {})
-  }
-};
